Generate Instrument model & migration
Schema difference compared to previous implementation:

When it comes to mapping customized keyboard bindings & persisting them,
instead of one column per keyboard key we are using the JSON data type.

This will not work with MySQL/MariaDb. It is a Postgres-specific data type.

As a result on the model level, it may change the interface with the
client, depending on how you wrap up the data as server responds to client.