refactor(GreenRoom) refactor render for CreateRoom
Simply move code into separate methods, as variables inside of the
regular render method. This highlights the rendering logic much more
easily and cleans up a lot of clutter. There may be a slight performance
hit, because things are encapsulated as functions now. But I think it
will not be noticable and the clarity gains are much more important.